+++ src/app/components/forms/dynamic-form/data/terra-form-field-input-number.ts
@@ -2,7 +2,7 @@ import {
     TerraFormFieldBase,
     TerraFormFieldBaseOptions
 } from './terra-form-field-base';
-import { ControlTypeEnum } from '../enum/controlType.enum';
+import { TerraControlTypeEnum } from '../enum/terra-control-type.enum';
 
 /**
  * @author mfrank
@@ -16,9 +16,9 @@ export class TerraFormFieldInputNumber extends TerraFormFieldBase<number>
 {
     public type:string;
 
-    constructor(key:string, options:TerraFormFieldInputNumberOptions = {})
+    constructor(key:string, label:string, required:boolean, options:TerraFormFieldInputNumberOptions = {})
     {
-        super(key, ControlTypeEnum.INPUT_NUMBER, options);
+        super(key, TerraControlTypeEnum.INPUT_NUMBER, label, required, options);
 
         this.type = options['type'] || '';
     }