+++ src/app/components/forms/dynamic-form/data/terra-form-field-input-text.ts
@@ -2,7 +2,7 @@ import {
     TerraFormFieldBase,
     TerraFormFieldBaseOptions
 } from './terra-form-field-base';
-import { ControlTypeEnum } from '../enum/controlType.enum';
+import { TerraControlTypeEnum } from '../enum/terra-control-type.enum';
 
 /**
  * @author mfrank
@@ -10,16 +10,25 @@ import { ControlTypeEnum } from '../enum/controlType.enum';
 export interface TerraFormFieldInputTextOptions extends TerraFormFieldBaseOptions<string>
 {
     type?:string;
+    isPassword?:boolean;
+    isIBAN?:boolean;
+    isReadOnly?:boolean;
 }
 
 export class TerraFormFieldInputText extends TerraFormFieldBase<string>
 {
     public type:string;
+    public isPassword:boolean;
+    public isIBAN:boolean;
+    public isReadOnly:boolean;
 
-    constructor(key:string, options:TerraFormFieldInputTextOptions = {})
+    constructor(key:string, label:string, required:boolean, options:TerraFormFieldInputTextOptions = {})
     {
-        super(key, ControlTypeEnum.INPUT_TEXT, options);
+        super(key, TerraControlTypeEnum.INPUT_TEXT, label, required, options);
 
-        this.type = options['type'] || '';
+        this.type = options.type || '';
+        this.isPassword = !!options.isPassword;
+        this.isIBAN = !!options.isIBAN;
+        this.isReadOnly = !!options.isReadOnly;
     }
 }