+++ src/app/components/forms/dynamic-form/service/terra-form-field-control.service.ts
@@ -10,6 +10,7 @@ import {
 import { isNullOrUndefined } from 'util';
 import { TerraFormFieldHorizontalContainer } from '../data/terra-form-field-horizontal-container';
 import { TerraFormFieldConditionalContainer } from '../data/terra-form-field-conditional-container';
+import { TerraFormFieldBaseContainer } from '../data/terra-form-field-base-container';
 
 /**
  * @author mfrank
@@ -21,7 +22,7 @@ export class TerraFormFieldControlService
     public defaultValues:{ [key:string]:string | number | boolean };
     public translationMapping:{ [key:string]:string };
 
-    private formFieldsToGroup:{ [key:string]:FormControl };
+    private formFieldsToGroup:{ [key:string]:any };
 
     constructor(private _formBuilder:FormBuilder)
     {
@@ -37,7 +38,7 @@ export class TerraFormFieldControlService
      */
     public createFormGroup(formFields:Array<TerraFormFieldBase<any>>):void
     {
-        this.initFormGroupHelper(formFields, false);
+        this.formFieldsToGroup = this.initFormGroupHelper(formFields, {}, false);
         this.dynamicFormGroup = this._formBuilder.group(this.formFieldsToGroup);
     }
 
@@ -49,41 +50,60 @@ export class TerraFormFieldControlService
         this.dynamicFormGroup.reset(this.defaultValues);
     }
 
+    /**
+     * Resets the form to default values
+     */
+    public updateDefaultValues(values:any):void
+    {
+        this.defaultValues = values;
+        this.resetForm();
+    }
+
     private initFormGroupHelper(formFields:Array<TerraFormFieldBase<any>>,
-                                isDisabled:boolean = false):void
+                                toGroup:{ [key:string]:any },
+                                isDisabled:boolean = false):{ [key:string]:any }
     {
         formFields.forEach((formField:TerraFormFieldBase<any>) =>
         {
-            if(formField instanceof TerraFormFieldHorizontalContainer && !isNullOrUndefined(formField.containerEntries))
+            if(formField instanceof TerraFormFieldBaseContainer && !isNullOrUndefined(formField.containerEntries))
             {
-                this.initFormGroupHelper(formField.containerEntries, false);
+                toGroup[formField.key] = this._formBuilder.group(this.initFormGroupHelper(formField.containerEntries, {}, false));
             }
             else if(formField instanceof TerraFormFieldConditionalContainer && !isNullOrUndefined(formField.conditionalEntries))
             {
-                this.formFieldsToGroup[formField.key] = new FormControl(formField.value, null);
-
-                for(let key in formField.conditionalEntries)
-                {
-                    if(formField.conditionalEntries.hasOwnProperty(key))
-                    {
-                        this.initFormGroupHelper(formField.conditionalEntries[key], true);
-                    }
-                }
+                // TODO extract into own component  or condition refactoring
+                //let subGroup:{ [key:string]:any } = {};
+                //
+                //subGroup[formField.key] = new FormControl(formField.value, this.generateValidators(formField));
+                //
+                //this.defaultValues[formField.key] = formField.value;
+                //
+                //for(let key in formField.conditionalEntries)
+                //{
+                //    if(formField.conditionalEntries.hasOwnProperty(key))
+                //    {
+                //        subGroup[key] = this._formBuilder.group(this.initFormGroupHelper(formField.conditionalEntries[key], {}, true));
+                //    }
+                //}
+                //
+                //toGroup[formField.key] = this._formBuilder.group(subGroup);
             }
             else
             {
-                this.formFieldsToGroup[formField.key] = new FormControl(formField.value, this.generateValidators(formField));
+                toGroup[formField.key] = new FormControl(formField.defaultValue, this.generateValidators(formField));
                 if(isDisabled)
                 {
-                    this.formFieldsToGroup[formField.key].disable({
+                    toGroup[formField.key].disable({
                         onlySelf:  true,
                         emitEvent: false
                     });
                 }
-                this.defaultValues[formField.key] = formField.value;
+                this.defaultValues[formField.key] = formField.defaultValue;
                 this.translationMapping[formField.key] = formField.label;
             }
         });
+
+        return toGroup;
     }
 
     private generateValidators(formField:TerraFormFieldBase<any>):Array<ValidatorFn>