+++ src/app/components/tables/data-table/context-menu/example/terra-data-table-context-menu.component.example.ts
@@ -1,0 +1,13 @@
+import {
+    Component
+} from "@angular/core";
+
+@Component({
+    selector: 'terra-data-table-context-menu-example',
+    styles:   [require('./terra-data-table-context-menu.component.example.scss')],
+    template: require('./terra-data-table-context-menu.component.example.html'),
+})
+export class TerraDataTableContextMenuComponentExample
+{
+
+}