+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 	- Refactoring -> Removed deprecated interface and input properties
 	- New `TerraDataTableBaseService` which replaces custom directives and configs. It also stores all information needed for retrieving data from the server
 
+### Feature
+* **url-helper** new component that offers methods to edit urls
+* **terra-multi-split-view** inject router instead of passing it via the input `inputRouter`.
 
 <a name="2.1.47-zlk"></a>
 # 2.1.47-zlk (06.02.2018)